% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{rclue-package}
\alias{rclue}
\alias{rclue-package}
\title{rclue: Expose the Student C++ Class using 'Rcpp Modules'}
\description{
Provides an implementation of a Student 'C++' class being connected to 'R' using 'Rcpp Modules' to expose the underlying class methods.
}
\details{
rclue is a package showing how to use Rcpp Modules to make available
a C++ class under S4.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-pkg-examples/rcpp-modules-student}
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-modules-student/issues}
}

}
\author{
\strong{Maintainer}: Jungwoo Chae \email{jungwoo@linewalks.com}

}
